board-style
I did some styling to the board and to the CSS to make it more pleasing to the eye with some text and 'Mine' and 'Flag' options. The board is not complete, as it is currently showing the form 'E' instead of a full board, so I will have to change that. I am not happy with the margins and positions of the board as of currently, so I need to fix that as well.